refactor: 학수번호 타입 수정
- Integer -> Long
- Wrapper 타입을 쓴 이유는 존재하지 않을 경우 null 값을 허용하기 위함

+++ src/main/java/com/gonghak98/v2/file/service/dto/FileData.java
@@ -1,6 +1,6 @@
 package com.gonghak98.v2.file.service.dto;
 
-public record FileData(int courseId,
+public record FileData(Long courseId,
                        String courseName,
                        int year,
                        int semester,
